apps/app-hija-1/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyJWT
pymysql
python-dotenv
mysql-connector-python
cryptography
